+++ apps/hpc-cdm/src/app/components/enketo/types.ts
@@ -1,0 +1,16 @@
+export type FormStatus =
+  | {
+      type: 'idle';
+    }
+  | {
+      type: 'saving';
+    }
+  | {
+      type: 'error';
+      message: string;
+    }
+  | {
+      type: 'conflict';
+      otherPerson: string;
+      timestamp: Date;
+    };
